OLPC ARM: Update strp so it gets a valid suspend power reading
Use the ec-sus-pwr command rather than ec-max-pwr for getting the
power when the host is in suspend.  ec-sus-pwr only takes a reading if
the EC thinks the host is in suspend.
Even with that a more delay was required for the time spent in
suspend than what wake1 was providing.  The EC only samples the gauge
every 100ms and the gauge itself only updates every 88ms.  Wake1 was
setting up a 1 second rtc wakeup but by the time the host actually
makes it into suspend the EC didn't have enough time left to get a
good consistent reading.
This patch makes a wake2 with a 2 second target and updates the
suspend test to use it.