Revert -mlittle-endian removal in r3596
The _real_ problem was user error.  I was trying to compile the arm
execuatable with an x86 compiler.  -mlittle-endian is a perfectly
valid option for the arm-elf-gcc compiler.  Thankfully its the default
so the removal didn't actually cause a problem.

Same sort of thing for the xinflate change in r3596 but I didn't
revert that becuase it does make it clearer whats going on and that
there are 2 versions of the inflate object.